Refactoring dto
