+++ src/main/kotlin/com/mda/diet/dto/Auth0SignupReturnDto.kt
@@ -1,5 +1,5 @@
-package com.mda.diet.model
+package com.mda.diet.dto
 
-data class Auth0SignupReturn(var _id: String = "",
+data class Auth0SignupReturnDto(var _id: String = "",
                              var email_verified: Boolean = false,
                              var email: String = "")
