+++ src/main/kotlin/com/mda/diet/dto/Auth0TokenReturnDto.kt
@@ -1,7 +1,6 @@
-package com.mda.diet.model
+package com.mda.diet.dto
 
-
-data class Auth0TokenReturn(var access_token: String = "",
+data class Auth0TokenReturnDto(var access_token: String = "",
                             var id_token: String = "",
                             var scope: String? = "",
                             var expires_in: Int = 0,
